Add support for switching between wired and wireless networks
Since this program already does everything necessary to switch between
networks, this is a fairly simple addition. Existing config files will
still work (the key "interface" is now an alias for
"wireless_interface"), but users now have the option of adding a
wired_interface and a network of type=wired, as in the updated
wireless.cfg example.